processor_attributes: split logic to prepare to support log metadata
The following patch splits the logic of the plugin that originally only
supports Traces attributes. Now certain functions has been moved to a
traces interface and some cleanup has been done in the attributes one.

This refactoring is still work in process.
