+++ plugins/processor_attributes/CMakeLists.txt
@@ -1,4 +1,6 @@
 set(src
-  attributes.c)
+  attributes.c
+  traces.c
+  variant_utils.c)
 
 FLB_PLUGIN(processor_attributes "${src}" "")
